% Auto-generated: do not edit by hand
\name{''PlayComponent}

\alias{''PlayComponent}

\title{PlayComponent component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''PlayComponent(id=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{value}{Logical. The value displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

